Updated user management module.

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,7 +6,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Models\InterestIn;
+use App\Models\UserInterestIn;
 use App\Models\OrganizationCategory;
 use App\Models\Users;
 use Illuminate\Http\Request;
@@ -20,8 +20,8 @@ class UserController extends Controller
     /** @var Users User model */
     protected $usersModel;
 
-    /** @var InterestIn InterestIn model */
-    protected $interestInModel;
+    /** @var UserInterestIn UserInterestIn model */
+    protected $userInterestInModel;
 
     /** @var OrganizationCategory OrganizationCategory model */
     protected $organizationCategoryModel;
@@ -30,13 +30,13 @@ class UserController extends Controller
      * Initialize UserController class.
      *
      * @param Users                $users                Users model
-     * @param InterestIn           $interestIn           InterestIn model
+     * @param UserInterestIn       $userInterestIn       UserInterestIn model
      * @param OrganizationCategory $organizationCategory OrganizationCategory model
      */
-    public function __construct( Users $users, InterestIn $interestIn, OrganizationCategory $organizationCategory )
+    public function __construct( Users $users, UserInterestIn $userInterestIn, OrganizationCategory $organizationCategory )
     {
         $this->usersModel                = $users;
-        $this->interestInModel           = $interestIn;
+        $this->userInterestInModel       = $userInterestIn;
         $this->organizationCategoryModel = $organizationCategory;
     }
 
@@ -57,17 +57,21 @@ class UserController extends Controller
     /**
      * Delete a specific user.
      *
-     * @param Request $request Request object
-     * @param Users   $users   Users model
+     * @param Request        $request        Request object
+     * @param Users          $users          Users model
+     * @param UserInterestIn $userInterestIn UserInterestIn model
      *
      * @return    \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse|\Illuminate\Http\JsonResponse
      */
-    public function destroy( Request $request, Users $user )
+    public function destroy( Request $request, Users $user, UserInterestIn $userInterestIn )
     {
-        $success = $user->delete();
+        $successDeleteUserInterest = $userInterestIn->where( [ 'fk_user_id' => $user->id ] )->delete();
+        if( $successDeleteUserInterest ){
+            $success = $user->delete();
+        }
 
         if( $request->ajax() ){
-            return response()->json( [ 'success' => $success ] );
+            return response()->json( [ 'success' => $success, 'message' => __('user_admin.user_management.remove_user_success'), 'redirectedUrl' => route( 'admin.user.index' ), ] );
         } else {
             return redirect()->route( 'admin.user.index' );
         }
